requirements.txt
python-telegram-bot
fastapi
uvicorn
openai
